requirements.txt
# Core utilities
numpy~=1.25.0
PyYAML~=6.0

# MTEB 벤치마크
mteb>=1.38.6

# Transformers ecosystem
transformers>=4.30.0
datasets>=2.14.0
sentence-transformers>=2.2.2

# (Optional) for caching & speed
torch>=2.0.0